+++ src/clippt/cli.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 import sys
 import shlex
+from typing import Callable
 
 import click
 
@@ -8,7 +9,7 @@ from clippt.app import PresentationApp
 from clippt.presentation import Presentation
 
 
-def common_options(func):
+def common_options(func: Callable) -> Callable:
     func = click.option("--no-footer", is_flag=True, help="Disable footer.")(func)
     func = click.option("--no-header", is_flag=True, help="Disable header.")(func)
     func = click.option(
@@ -16,6 +17,7 @@ def common_options(func):
     )(func)
     func = click.option("--serve", "-s", is_flag=True, help="Start a web server")(func)
     func = click.option("-v", "--verbose", count=True)(func)
+    func = click.option("--theme", "-t", help="Theme to select")(func)
     return func
 
 
@@ -34,21 +36,46 @@ def clippt(*, source: Path, verbose: int, **kwargs):
     """Run a presentation in the command-line."""
     _apply_log_level(verbose)
     presentation = Presentation.from_path(source)
-    run_cli(
+    _run_cli(
         presentation=presentation,
         **kwargs,
     )
 
 
-def run_cli(
+def create_cli_command(presentation: Presentation):
+    """Create a CLI command for a concrete presentation.
+
+    Useful when using clippt as a library."""
+
+    def app(*, verbose: int, **kwargs):
+        _apply_log_level(verbose)
+        _run_cli(
+            presentation=presentation,
+            **kwargs,
+        )
+
+    app.__doc__ = (
+        f"""Present '{presentation.title}'."""
+        if presentation.title
+        else "Run the presentation"
+    )
+    app = common_options(app)
+    app = click.command()(app)
+    return app
+
+
+def _run_cli(
     *,
     presentation: Presentation,
     no_footer: bool,
     no_header: bool,
     continue_: bool,
+    theme: str | None,
     serve: bool,
 ):
     app = PresentationApp(presentation=presentation)
+    if theme:
+        app.theme = theme
     app.enable_footer = not no_footer
     app.enable_header = not no_header
     if continue_ and Path(".current_slide").exists():
@@ -77,28 +104,6 @@ def run_cli(
         server.serve()
     else:
         app.run()
-
-
-def create_cli_command(presentation: Presentation):
-    """Create a CLI command for a concrete presentation.
-
-    Useful when using clippt as a library."""
-
-    def app(*, verbose: int, **kwargs):
-        _apply_log_level(verbose)
-        run_cli(
-            presentation=presentation,
-            **kwargs,
-        )
-
-    app.__doc__ = (
-        f"""Present '{presentation.title}'."""
-        if presentation.title
-        else "Run the presentation"
-    )
-    app = common_options(app)
-    app = click.command()(app)
-    return app
 
 
 def _apply_log_level(verbose: int) -> None: