+++ src/clippt/theming.py
@@ -1,26 +1,10 @@
-from textual.theme import Theme
+"""Theming and styling for clippt.
 
-my_theme = Theme(
-    name="my",
-    primary="#0000c0",
-    secondary="#4040ff",
-    accent="#00ff00",
-    foreground="#444444",
-    background="#ffffff",
-    success="#A3BE8C",
-    warning="#EBCB8B",
-    error="#BF616A",
-    surface="#ffffff",
-    panel="#ffffff",
-    dark=False,
-    variables={
-        "block-cursor-text-style": "none",
-        "footer-key-foreground": "#88C0D0",
-        "input-selection-background": "#81a1c1 35%",
-    },
-)
+Useful links:
+- https://textual.textualize.io/widgets/markdown/#textual.widgets._markdown.Markdown.BLOCKS
+"""
 
-css_tweaks = """
+css_tweaks: str = """
     Screen {
         align: center middle;
     }
@@ -34,10 +18,10 @@ css_tweaks = """
     Header {
         height: 1;
     }
-    Markdown.title {
-        MarkdownFence {
-            margin: 0;
-        }
+    Markdown {
+      MarkdownH1 {
+        margin: 1 0;
+      }
     }
     Static.slide-title {
         height: 1;
@@ -59,9 +43,9 @@ css_tweaks = """
         max-height: 500;
     }
     Static.output {
-        margin: 0 3;
+        margin: 0 0;
         max-height: 500;
-        padding: 1 2;
+        padding: 1 1;
     }
     Static.error {
         background: #ffcccc;
@@ -70,4 +54,5 @@ css_tweaks = """
         max-height: 500;
         padding: 1 2;
     }
- """
+"""
+"""Additional CSS styling for the presentation app."""